LayoutTests/fast/scrolling/scroll-animator-select-list-events-expected.txt

CONSOLE MESSAGE: line 16: MainFrameView: mouseEnteredContentArea
CONSOLE MESSAGE: line 16: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 17: RenderListBox: mouseEnteredContentArea
CONSOLE MESSAGE: line 17: RenderListBox: mouseMovedInContentArea
CONSOLE MESSAGE: line 17: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 18: RenderListBox: mouseEnteredVerticalScrollbar
CONSOLE MESSAGE: line 18: RenderListBox: mouseMovedInContentArea
CONSOLE MESSAGE: line 18: MainFrameView: mouseMovedInContentArea
CONSOLE MESSAGE: line 19: RenderListBox: mouseIsDownInVerticalScrollbar
CONSOLE MESSAGE: line 20: RenderListBox: mouseIsUpInVerticalScrollbar
CONSOLE MESSAGE: line 21: RenderListBox: mouseExitedVerticalScrollbar
CONSOLE MESSAGE: line 21: RenderListBox: mouseExitedContentArea
CONSOLE MESSAGE: line 21: MainFrameView: mouseMovedInContentArea
PASS successfullyParsed is true

TEST COMPLETE
Test for https://bugs.webkit.org/show_bug.cgi?id=153398.
